requirements.txt
# Core Libraries
Flask==2.3.2  # Web framework for building the application
Werkzeug==2.3.2  # WSGI utility library used by Flask
python-dotenv==1.0.0  # For loading environment variables from .env file

# Data Processing
pandas==1.5.3  # For handling and processing CSV files
numpy==1.24.3  # For numerical operations

# Machine Learning
scikit-learn==1.2.2  # For clustering, classification, and evaluation
sentence-transformers==2.2.2  # For generating embeddings using pre-trained models
joblib==1.3.1  # For saving and loading machine learning models

# LLM Integration
groq==1.0.0  # For interacting with the Groq API for LLM-based classification

# Environment and API
# The .env file contains the GROQ_API_KEY required for accessing the Groq API.
# Ensure the .env file is properly configured with the following:
# GROQ_API_KEY="your_groq_api_key_here"
